+++ selendroid-standalone/src/main/java/io/selendroid/SelendroidConfiguration.java
@@ -27,9 +27,6 @@ public class SelendroidConfiguration {
   @Parameter(description = "timeout that will be used to start Android emulators", names = "-timeoutEmulatorStart")
   private long timeoutEmulatorStart = 300000;
 
-  @Parameter(description = "if true, adb will be restarted while starting selendroid-standalone.", names = "-restartAdb")
-  private boolean restartAdb = false;
-
   @Parameter(description = "location of the application under test. Absolute path to the apk", names = {
       "-app", "-aut" })
   private List<String> supportedApps = new ArrayList<String>();
@@ -130,14 +127,6 @@ public class SelendroidConfiguration {
 
   public void setTimeoutEmulatorStart(long timeoutEmulatorStart) {
     this.timeoutEmulatorStart = timeoutEmulatorStart;
-  }
-
-  public boolean isRestartAdb() {
-    return restartAdb;
-  }
-
-  public void setRestartAdb(boolean restartAdb) {
-    this.restartAdb = restartAdb;
   }
 
   public boolean isVerbose() {